feat: implement PredictionNameInput component and integrate it into SelectDatasetStep and SelectModelStep
